requirements.txt
Appium-Python-Client==4.0.0
pytest==8.2.2
